requirements.txt
fastapi==0.111.0
pydantic==2.7.1
pydantic_settings==2.2.1
PyJWT==2.8.0
pytest==8.1.2
python-decouple==3.8
bcrypt==4.1.3
SQLAlchemy==2.0.29
starlette==0.37.2
uvicorn==0.29.0
